# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients/{patientId}/invitations:
    post:
      tags: [recovery]
      operationId: createAccessInvitation
      summary: Create a one-time patient access invitation
      parameters:
        - $ref: "#/components/parameters/PatientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationInput"
      responses:
        "201":
          description: Invitation created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResult"
  /patients:
    get:
      operationId: listPatients
      tags: [recovery]
      summary: List patients
      responses:
        "200":
          description: Patients assigned to the current clinician
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [recovery]
      summary: Add a patient and assign an initial exercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created patient and shareable invite code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePatientResult"
  /patients/{patientId}:
    get:
      operationId: getPatient
      tags: [recovery]
      summary: Get a patient
      parameters:
        - $ref: "#/components/parameters/PatientId"
      responses:
        "200":
          description: Patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Patient not found
  /patients/{patientId}/trend:
    get:
      operationId: getPatientTrend
      tags: [recovery]
      summary: Get patient recovery trends
      parameters:
        - $ref: "#/components/parameters/PatientId"
      responses:
        "200":
          description: Trend data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientTrend"
  /sessions:
    get:
      operationId: listSessions
      tags: [recovery]
      summary: List recent sessions
      parameters:
        - name: patientId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Recent sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
    post:
      operationId: createSession
      tags: [recovery]
      summary: Save a completed exercise session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionInput"
      responses:
        "201":
          description: Saved session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
  /programs:
    get:
      operationId: listPrograms
      tags: [recovery]
      summary: List exercise programs
      parameters:
        - name: patientId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Exercise programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExerciseProgram"
    post:
      operationId: createProgram
      tags: [recovery]
      summary: Create an exercise program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramInput"
      responses:
        "201":
          description: Exercise program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseProgram"
  /programs/{programId}:
    patch:
      operationId: updateProgram
      tags: [recovery]
      summary: Update an exercise program
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramInput"
      responses:
        "200":
          description: Updated exercise program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseProgram"
...
  /alerts/{alertId}/review:
    patch:
      operationId: reviewAlert
      tags: [recovery]
      summary: Mark an alert as reviewed
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reviewed alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /sports/screenings:
    get:
      operationId: listScreenings
      tags: [sports]
      summary: List recent squad screenings
      responses:
        "200":
          description: Sports screening results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Screening"
    post:
      operationId: createScreening
      tags: [sports]
      summary: Save a sports screening
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreeningInput"
      responses:
        "201":
          description: Screening result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Screening"
components:
  parameters:
    PatientId:
      name: patientId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    CreateInvitationInput:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [patient, caregiver, coach] }
    InvitationResult:
      type: object
      required: [code, patientId, role, expiresAt]
      properties:
        code: { type: string }
        patientId: { type: string }
        role: { type: string, enum: [patient, caregiver, coach] }
        expiresAt: { type: string }
...
    Patient:
      type: object
      required: [id, name, contact, condition, assignedClinician, recoveryScore, adherence, hasAlert]
      properties:
        id: { type: string }
        name: { type: string }
        initials: { type: string }
        contact: { type: string }
        condition: { type: string }
        assignedClinician: { type: string }
        sport: { type: ["string", "null"] }
        recoveryScore: { type: number }
        adherence: { type: number }
        hasAlert: { type: boolean }
        lastSession: { type: ["string", "null"] }
    PatientInput:
      type: object
      required: [name, contact, condition, initialExercise]
      properties:
        name: { type: string, minLength: 2 }
        contact: { type: string, minLength: 3 }
        condition: { type: string, minLength: 2 }
        initialExercise: { type: string, minLength: 2 }
        sport: { type: string }
    CreatePatientResult:
      type: object
      required: [patient, inviteCode]
      properties:
        patient: { $ref: "#/components/schemas/Patient" }
        inviteCode: { type: string }
    Session:
      type: object
      required: [id, patientId, exercise, timestamp, totalReps, correctReps, averageAngle, movementScore, torsoAlignment, symmetry, painScore, adherence, mode, recommendation]
      properties:
        id: { type: string }
        patientId: { type: string }
        exercise: { type: string }
        timestamp: { type: string }
        totalReps: { type: number }
        correctReps: { type: number }
        averageAngle: { type: number }
        movementScore: { type: number }
        torsoAlignment: { type: number }
        symmetry: { type: number }
        painScore: { type: number, minimum: 0, maximum: 10 }
        adherence: { type: number }
        mode: { type: string, enum: [self, caregiver-assisted] }
        recommendation: { type: string }
    SessionInput:
      type: object
      required: [patientId, exercise, totalReps, correctReps, averageAngle, movementScore, torsoAlignment, symmetry, painScore, mode, protocolMode, recommendation]
      properties:
        patientId: { type: string }
        exercise: { type: string }
        totalReps: { type: number }
        correctReps: { type: number }
        averageAngle: { type: number }
        movementScore: { type: number }
        torsoAlignment: { type: number }
        symmetry: { type: number }
        painScore: { type: number, minimum: 0, maximum: 10 }
        mode: { type: string, enum: [self, caregiver-assisted] }
        protocolMode: { type: string, enum: [patient, pilot-demo] }
        recommendation: { type: string }
    TrendPoint:
      type: object
      required: [date, movementScore, averageAngle, painScore, adherence]
      properties:
        date: { type: string }
        movementScore: { type: number }
        averageAngle: { type: number }
        painScore: { type: number }
        adherence: { type: number }
    PatientTrend:
      type: object
      required: [patientId, points, summary]
      properties:
        patientId: { type: string }
        points:
          type: array
          items: { $ref: "#/components/schemas/TrendPoint" }
        summary: { type: string }
    ExerciseProgram:
      type: object
      required: [id, patientId, exercise, targetReps, targetSets, targetAngle, frequency]
      properties:
        id: { type: string }
        patientId: { type: string }
        exercise: { type: string }
        targetReps: { type: number }
        targetSets: { type: number }
        targetAngle: { type: number }
        frequency: { type: string }
    ProgramInput:
      type: object
      required: [patientId, exercise, targetReps, targetSets, targetAngle, frequency]
      properties:
        patientId: { type: string }
        exercise: { type: string }
        targetReps: { type: number }
        targetSets: { type: number }
        targetAngle: { type: number }
        frequency: { type: string }
    Alert:
      type: object
      required: [id, patientId, type, message, createdAt, resolved]
      properties:
        id: { type: string }
        patientId: { type: string }
        type: { type: string }
        message: { type: string }
        createdAt: { type: string }
        resolved: { type: boolean }
    Screening:
      type: object
      required: [id, athlete, sport, timestamp, score, status, finding, recommendation]
      properties:
        id: { type: string }
        patientId: { type: ["string", "null"] }
        athlete: { type: string }
        sport: { type: string }
        timestamp: { type: string }
        score: { type: number }
        status: { type: string, enum: [CLEAR, REVIEW] }
        finding: { type: string }
        recommendation: { type: string }
    ScreeningInput:
      type: object
      required: [athlete, sport, score, status, finding, recommendation]
      properties:
        patientId: { type: string }
        athlete: { type: string }
        sport: { type: string }
        score: { type: number }
        status: { type: string, enum: [CLEAR, REVIEW] }
        finding: { type: string }
        recommendation: { type: string }