# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Alert {
  id: string;
  patientId: string;
  type: string;
  message: string;
  createdAt: string;
  resolved: boolean;
}