# lib/api-zod/src/generated/types/createPatientResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Patient } from './patient';

export interface CreatePatientResult {
  patient: Patient;
  inviteCode: string;
}