# lib/api-zod/src/generated/types/invitationResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InvitationResultRole } from './invitationResultRole';

export interface InvitationResult {
  code: string;
  patientId: string;
  role: InvitationResultRole;
  expiresAt: string;
}