# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: string;
  name: string;
  initials?: string;
  contact: string;
  condition: string;
  assignedClinician: string;
  /** @nullable */
  sport?: string | null;
  recoveryScore: number;
  adherence: number;
  hasAlert: boolean;
  /** @nullable */
  lastSession?: string | null;
}