# lib/api-zod/src/generated/types/patientTrend.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrendPoint } from './trendPoint';

export interface PatientTrend {
  patientId: string;
  points: TrendPoint[];
  summary: string;
}