# lib/api-zod/src/generated/types/programInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProgramInput {
  patientId: string;
  exercise: string;
  targetReps: number;
  targetSets: number;
  targetAngle: number;
  frequency: string;
}