# lib/api-zod/src/generated/types/screening.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScreeningStatus } from './screeningStatus';

export interface Screening {
  id: string;
  /** @nullable */
  patientId?: string | null;
  athlete: string;
  sport: string;
  timestamp: string;
  score: number;
  status: ScreeningStatus;
  finding: string;
  recommendation: string;
}